requirements.txt
streamlit
audio-recorder-streamlit
scipy
requests
numpy
python-dotenv==1.0.0
groq
